+++ engine/launcher/scripts/build.sh
@@ -3,23 +3,18 @@
 
 set -euo pipefail   # Bash "strict mode"
 script_dirpath="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
-root_dirpath="$(dirname "${script_dirpath}")"
+lang_root_dirpath="$(dirname "${script_dirpath}")"
 
 
 
 # ==================================================================================================
 #                                             Constants
 # ==================================================================================================
-RELEASE_SCRIPT_FILENAME="release-repo.sh"     # NOTE: Must be on the path; comes from devtools repo
-
-UPDATE_OWN_VERSION_CONSTS_SCRIPT_FILENAME="update-own-version-constants.sh"
-UPDATE_PACKAGE_VERSION_SCRIPT_FILENAME="update-package-versions.sh"
 
 
 # ==================================================================================================
 #                                             Main Logic
 # ==================================================================================================
-if ! bash "${RELEASE_SCRIPT_FILENAME}" "${root_dirpath}" "${script_dirpath}/${UPDATE_OWN_VERSION_CONSTS_SCRIPT_FILENAME}" "${script_dirpath}/${UPDATE_PACKAGE_VERSION_SCRIPT_FILENAME}"; then
-    echo "Error: Couldn't cut the release" >&2
-    exit 1
-fi
+cd "${lang_root_dirpath}"
+go test ./...
+go build ./...