+++ engine/scripts/build.sh
@@ -1,56 +1,29 @@
 #!/usr/bin/env bash
-set -euo pipefail # Bash "strict mode"
+# 2021-07-08 WATERMARK, DO NOT REMOVE - This script was generated from the Kurtosis Bash script template
+
+set -euo pipefail   # Bash "strict mode"
 script_dirpath="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 root_dirpath="$(dirname "${script_dirpath}")"
+
+
 
 # ==================================================================================================
 #                                             Constants
 # ==================================================================================================
-source "${script_dirpath}/_constants.env"
+BUILD_SCRIPT_RELATIVE_FILEPATHS=(
+    "api/golang/scripts/build.sh"
+    "api/typescript/scripts/build.sh"
+    "server/scripts/build.sh"
+)
 
-BUILD_DIRNAME="build"
-ENGINE_DIRNAME="engine"
-MAIN_BINARY_OUTPUT_FILE="kurtosis-engine.bin"
-MAIN_BINARY_OUTPUT_PATH="${root_dirpath}/${BUILD_DIRNAME}/${MAIN_BINARY_OUTPUT_FILE}"
-MAIN_GO_FILEPATH="${root_dirpath}/${ENGINE_DIRNAME}/main.go"
 
-# =============================================================================
-#                                 Main Code
-# =============================================================================
-# Checks if dockerignore file is in the root path
-if ! [ -f "${root_dirpath}"/.dockerignore ]; then
-  echo "Error: No .dockerignore file found in root '${root_dirpath}'; this is required so Docker caching is enabled and the engine image builds remain quick" >&2
-  exit 1
-fi
-
-# Test code
-echo "Running unit tests..."
-if ! go test "${root_dirpath}/..."; then
-  echo "Tests failed!" >&2
-  exit 1
-fi
-echo "Tests succeeded"
-
-# Build binary for packaging inside an Alpine Linux image
-echo "Building Kurtosis Engine Server code '${MAIN_GO_FILEPATH}'..."
-if ! CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o "${MAIN_BINARY_OUTPUT_PATH}" "${MAIN_GO_FILEPATH}"; then
-  echo "Error: Code build of the Kurtosis Engine Server failed" >&2
-  exit 1
-fi
-
-# Generate Docker image tag
-get_docker_image_tag_script_filepath="${script_dirpath}/${GET_DOCKER_IMAGE_TAG_SCRIPT_FILENAME}"
-if ! docker_tag="$(bash "${get_docker_image_tag_script_filepath}")"; then
-    echo "Error: Couldn't get the Docker image tag" >&2
-    exit 1
-fi
-
-# Build Docker image
-dockerfile_filepath="${root_dirpath}/${ENGINE_DIRNAME}/Dockerfile"
-image_name="${IMAGE_ORG_AND_REPO}:${docker_tag}"
-echo "Building Kurtosis Engine Server into a Docker image named '${image_name}'..."
-if ! docker build -t "${image_name}" -f "${dockerfile_filepath}" "${root_dirpath}"; then
-  echo "Error: Docker build of the Kurtosis Engine Server failed" >&2
-  exit 1
-fi
-echo "Successfully built Docker image '${image_name}' containing the Kurtosis Engine Server"
+# ==================================================================================================
+#                                             Main Logic
+# ==================================================================================================
+for build_script_rel_filepath in "${BUILD_SCRIPT_RELATIVE_FILEPATHS[@]}"; do
+    build_script_abs_filepath="${root_dirpath}/${build_script_rel_filepath}"
+    if ! bash "${build_script_abs_filepath}"; then
+        echo "Error: Build script '${build_script_abs_filepath}' failed" >&2
+        exit 1
+    fi
+done