+++ engine/server/engine/enclave_manager/enclave_manager.go
@@ -7,8 +7,8 @@ import (
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager"
 	"github.com/kurtosis-tech/container-engine-lib/lib/docker_manager/types"
 	"github.com/kurtosis-tech/kurtosis-core/launcher/api_container_launcher"
-	"github.com/kurtosis-tech/kurtosis-engine-api-lib/golang/kurtosis_engine_rpc_api_bindings"
-	"github.com/kurtosis-tech/kurtosis-engine-server/engine/enclave_manager/docker_network_allocator"
+	"github.com/kurtosis-tech/kurtosis-engine-server/api/golang/kurtosis_engine_rpc_api_bindings"
+	"github.com/kurtosis-tech/kurtosis-engine-server/server/engine/enclave_manager/docker_network_allocator"
 	"github.com/kurtosis-tech/object-attributes-schema-lib/forever_constants"
 	"github.com/kurtosis-tech/object-attributes-schema-lib/schema"
 	"github.com/palantir/stacktrace"
@@ -20,9 +20,6 @@ import (
 	"sync"
 )
 const (
-	// This is set in the API container Dockerfile
-	availabilityWaiterBinaryFilepath = "/run/api-container-availability-waiter"
-
 	shouldFetchStoppedContainersWhenGettingEnclaveStatus = true
 
 	// We set this to true in case there are any race conditions with a container starting as we're trying to stop the enclave